backend/requirements.txt
Django>=4.2
djangorestframework
channels
channels_redis
daphne
google-generativeai
dj_database_url
python-dotenv
djangorestframework-simplejwt
PyJWT
django-cors-headers
pymongo
mongoengine